# backend/src/api/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Poker-API
  description: Progrendszerek projekt backend apija (Poker oldal no money)
...
paths:
...
  /games:
    get:
      summary: Get all games by query that user has access to
      operationId: getGames
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: myGames
          schema:
            type: boolean
          required: false
          description: If true, only games where the user is a player will be returned
        - in: query
          name: name
          schema:
            type: string
            nullable: true
          allowEmptyValue: true
          required: false
          description: The name of the game
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: The page number to retrieve
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GamesResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
  /game:
    delete:
      summary: Delete game
      operationId: deleteGame
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: gameId
          schema:
            type: string
          required: true
          description: The id of the game to delete
      responses:
        '200':
          description: Game deleted successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - gameOwnerAuth: []
        - cookieAdminAuth: []
    put:
      summary: Update game
      operationId: updateGame
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: gameId
          schema:
            type: string
          required: true
          description: The id of the game to retrieve
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId:
                  type: string
                options:
                  type: object
                  properties:
                    key:
                      type: string
                    whiteList:
                      type: array
                      items:
                        type: string
                    banList:
                      type: array
                      items:
                        type: string
      responses:
        '200':
          description: Game updated successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - gameOwnerAuth: []
        - cookieAdminAuth: []
    get:
      summary: Get game by id
      operationId: getGameById
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: gameId
          schema:
            type: string
          required: true
          description: The id of the game to retrieve
      responses:
        '200':
          description: Game retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - gameOwnerAuth: []
        - cookieAdminAuth: []
    post:
      summary: Create a new game
      operationId: createGame
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  type: object
                  properties:
                    whiteList:
                      type: array
                      items:
                        type: string
                      description: The ids of the users who can join the game or usernames
                    maxPlayers:
                      type: number
                      description: The maximum number of players in the game
                    isPublic:
                      type: boolean
                      description: If true, anyone can join the game
                    maxRaises:
                      type: number
                      description: The maximum number of raises in a round
              required:
                - options
      responses:
        '200':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
  /game/invite:
    post:
      summary: Invite user to game
      operationId: inviteUser
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: false
          description: The id of the user to invite
        - in: query
          name: gameId
          schema:
            type: string
          required: true
          description: The id of the game to invite to
      responses:
        '200':
          description: User invited successfully
        '404':
          description: User or game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - gameOwnerAuth: []
        - cookieAdminAuth: []
  /game/join:
    get:
      summary: Join game
      operationId: joinGame
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: token
          schema:
            type: string
          required: false
          description: The magic token to join the game with
        - in: query
          name: gameId
          schema:
            type: string
          required: false
          description: The id of the game to join
      responses:
        '200':
          description: Game joined successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - cookieAuth: []
  /game/leave:
    put:
      summary: Leave game
      operationId: leaveGame
      x-eov-operation-handler: api/controllers/game.controller
      tags:
        - Game
      parameters:
        - in: query
          name: gameId
          schema:
            type: string
          required: true
          description: The id of the game to leave
      responses:
        '200':
          description: Game left successfully
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseError"
      security:
        - gamePlayerAuth: []
# Chat
...
components:
  schemas:
...
    GamesResponse: # TODO: remove unvanted fields like user balance, and populate user, also remove cards etc.
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
          ownerId:
            type: string
          chatChannelId:
            type: string
          players:
            type: array
            items:
              type: object
              properties:
                userId:
                  type: string
                positionAtTable:
                  type: number
                stillPlayingInRound:
                  type: boolean
                leftGame:
                  type: boolean
          round:
            type: number
          phase:
            type: string
    GameResponse: # TODO: remove unvanted fields like user balance, and populate user
      type: object
      properties:
        _id:
          type: string
        ownerId:
          type: string
        chatChannelId:
          type: string
        players:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              cards:
                type: array
                items:
                  type: object
                  properties:
                    value:
                      type: string
                    sign:
                      type: string
              inGameBalance:
                type: number
              bet:
                type: number
              called:
                type: boolean
              raisedTimes:
                type: number
              positionAtTable:
                type: number
              stillPlayingInRound:
                type: boolean
              leftGame:
                type: boolean
        cardsOnTable:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              sign:
                type: string
        cardsInDeck:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              sign:
                type: string
        round:
          type: number
        paused:
          type: boolean
        phase:
          type: string
        options:
          type: object
          properties:
            key:
              type: string
            whiteList:
              type: array
              items:
                type: string
            banList:
              type: array
              items:
                type: string
...
    BaseError:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        status:
          type: number
        originatedFrom:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    userSpecificCookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    chatMemberAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gameOwnerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    gamePlayerAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    cookieAdminAuth:
      type: apiKey
      in: cookie
      name: connect.sid

# TODO: mindehol jo response
# TODO: error messagek ahol lehet
# TODO: example es description
# TODO: plural or not plural, also generic all endpoint or single endpoint?
# FIXME: change query to path param where needed